[SOW 3] Fix race conditions in job startup (#1549)
* Make runner_test.py support both setup.py test and direct pytest

* Add gc job to clean up any stale temporary directories we might leak

* Fix race conditions in job startup

The flask server no longer writes to run_dir directly during job
startup; instead, it writes to a temporary directory, and then passes
ownership of the temporary directory to the runner daemon. This is the
first step towards the flask server never touching the run_dir
directly.

Under normal circumstances, the temp dir is always cleaned up
immediately after creation. In case of abnormal circumstances, the
runner daemon now checks once a day for any stale temp dirs, and
cleans them up.

The runner daemon is now careful to make sure that you never have
multiple jobs running in the same run_dir, or old jobs running in a
new run_dir.

* refactor runner daemon process handling into a separate file

In preparation for adding docker support

* First pass docker-based backend

* Make docker backend selectable by feature flag

* Refactor process cleanup logic to fix a small race condition

* Add a bit more logging to docker_process.py

* Let server.py select backend on a job-by-job basis

* Rename job -> report_job everywhere, to prepare for extract jobs

* Fix pass runner support for extract jobs

* First attempt at plumbing extract jobs throughout, doesn't work

* prepare_output_file now takes run_dir instead of report_info

- Computing report_info requires access to the flask session and all
  kinds of things, which isn't suitable for running in a remote
  extraction job.

- The only thing any of the prepare_output_file implementations use
  report_info for is to get the run_dir.

* don't crash on old run_dir's

* Remove pkdp

* use trailing commas more consistently

* Fix background_percent_complete extract job

* In docker backend, use a different working directory

Just to make sure that the server isn't directly poking around in
run_dir, and that we're using extract jobs for everything we need to.

* Address review comments

* Rename _params to _input_params