Do not delete self from instances.
Driver's persist for the life of the system so they should never be removed from
instances. In the event that the websocket connection is broken with the agent
(ex agent dies unexpectedly) then there is no reason to remove self from
instances. The agent will be re-started on the next request.