Fix #2518. Deleted run dirs do not result in an error.
There are a few fixes:
1. In __db_init we use prev_db to get the jobRunMode.
2. When a run_dir is purged we initiliaze a new supervisor db.
Previously we didn't initialize a new db so the gui thought
there were frames available and requested them resulting in
"Report not generated." Initializing the db solves this by
re-initializing parallelStatus so the frameCount is 0 and no frames
are requested.
3. If a job status is in job_run_purged we need to add it to the
purged_jids_cache. Without this we read the db file for purged jobs
each time purge_free_simulations() runs.