Fix #2606. Control when auth_db is initialized.
This partially reverts pulls/2594. In that pull we moved
auth_db.init() to the module level and the auth_db was initialized on
import. This caused agents to init the db which they don't have access
to. This reverts part of that pull so auth_db.init() is called by
those who need it initialized.