Remove checks for isLoaded() in simHandleStatus, they are unnecessary.
simHandleStatus will only ever be called when the models are loaded.
- simHandleStatus is registered in initSimulationState
- in initSimulationState simHandleStatus is called in handleStatus
which is called in runStatus
- runStatus is called whenModelsLoaded() so checking it in
simHandleStatus is redundant

There is one edge case, when we leave the page but there is still
a runStatus request pending. In that case on $destroy we clearSimulation
(unload the models) and set the q item (pending runStatus request)
state to 'removing'. When the request finally resolves we check
qi.state == 'removing' and don't call the callback (handleStatus)
if this is the case. So again, the check for isLoaded is redundant.