Fix #2712. Make waiting on websocket_ready ordered.
Tornado Events use a set() under the hood. Because sets are
unordered when the waiters are woken up the waking happens in
a random order. We uses tornado Events to wait for the agent
websocket to be ready. We need the waiters on the websocket
to be released in an ordered fashion. Not doing so was causing
undesirable behavior. Ex if we have a [run, cancel, run] waiting on
the websocket. When they are released maybe they then execute
in [run, run, cancel] order. This won't work because cancel's
should only cancel the ops ahead of it. Using an orderedSet enusres
that the waiters are woken up in the order they came in.