Fix #2934. Remove "simulation" from ML app.
Simulation was removed from ML app. In addition, a refactor to
move more strings onto the schema and to create consistency in messaging.
For example, in the "Welcome text" when the link to add a new simulation
says "New Dataset" then the welcome text will also say "New Dataset".
Previosly it would have said "New Simulation".