Fix #3145. Do not evaluate RPN's that are shell commands.
RPN namelist fields in elegant can also be strings containing shell commands
that will be evaluated by elegant. In this case just pass the string through
and do not try to evaluate it.

In the sirepo.lib case this is fine and will always work. In the Sirepo GUI
case it may not always work.

There are many possible failure points for these types of commands. For example,
when converting from elegant to madx some RPN fields we must evaluate. If the
field contains shell commands then we will get the string of the commands and
may try to do math operations on it. We will fail at that point. Another example,
is that we mangle filenames but don't mangle them in the shell commands. So, if
a user has shell commands with filenames in them the filenames will not be found
when elegant evaluates the command.