WIP radiasoft/sirepoflash#16 allow user to upload FLASH problem files
Users can now upload a ZIP archive containing FLASH problem files
(F90, Config, and Makefile). These will be unpacked in the proper location
in the FLASH source code and used when compiling.

Some notable changes:
- The FLASH source code is no longer distributed as an rpm. It is
distributed as a tarball that lives in the user's lib dir.
- CapLaser3D and CapLaserBELLA are no longer distributed inside
of the FLASH source code. They are now files in the lib dir. They are
distributed alongside the FLASH source code. Upon getting the FLASH role
they are unpacked in the user's lib dir.