Fix #3350: wrap calls to flask.g
Inside of a flask request should be the only time we touch
flask.g.  We need to use flask.g here because flask is threaded
and flask.g handles getting the right object for the thread for
us. Outside of flask (tornado, jupyterhub) we can use a global
variable because it is single-threaded. All calls must be wrapped
in a `with` to ensure we maintain the thread of control (no
`await`'s).