Remove ability to edit setupConfigDirectives and runtime parameters
After some dicussion we've decided that having these fields be
editable doesn't make sense. They overlap with the fields offered in
the Source and Physics tabs which is confusing. For now remove this
feature but leave the Config and Runtime Parameter tabs as just static
tabs that reflect the information but don't allow the user to change
it.

There is some thought that this feature may be useful in the future.
In that case this commit could be reverted.

Known bugs: - When changing the type of directive_PARAMETER in the
'edit directive' modal the defualt and value fields need to be cleared
out. There is a TODO in the code to handle this. It is tricky because
it is hard to get notified when only the field has been changed (not
the whole model changing).