Fix #3505: Put sqlalchemy session on srcontext
In #3505 the sqlalchemy session was not rolled back after a failed
commit. We were using a global session across the app which brought
Sirepo down.

This puts the sqlalchemy session on srcontext. The seesion is
instantiated and destroyed/rolled back with each request (in
uri_router._dispatch).

Not using a global session means that all code that needs to use the
auth_db needs to have an instantiated session. In Flask it is created
and destroyed with each request (_dispatch). For pkcli, tests, and in
the supervisor we must create the context by hand everywhere we need
it.

For pkcli we considered creating the context in
sirepo_console.py. This works fine in Flask where `sirepo service
http` ends up creating a new process so the context in
sirepo_console.py doesn't conflict with the one created in _dispatch
But, in the case of the job_supervisor we are in the same process as
the "spawning" console process. srcontext is not reentrant so we would
have to use one global context for the whole supervisor. This would
mean if the supervisor had one failed commit then the session is
ruined for the whole process which is a problem.