Make sure to cancel "open" run requests
Run requests are held open until the run actually starts on the agent.
In the event that there is an error in test_http the run request is
already sent to Sirepo but then the request is closed by test_http.
This leaves an "orphaned" run that is running in Sirepo (or waiting to be run)
but test_http isn't monitoring. So as soon as the run is sent to Sirepo
no matter what happens we need to cancel it.