WIP #2598: use sim_db_file api for Radia sim files.
This is a very rough hacked up version of showing how to use the
sim_db_file api in radia.  The basic idea is once we create the file we
put it back to the server then when the animation starts on sbatch the
file will be requested and place in the run_dir.

There is some confusion about run_dir in the radia template
code. Sometimes run_dir has the same meaning as run_dir elsewhwere in
sirepo; an abspath to a dir where a compute happened. Other times it
just means "basename of a dir where there are some file(s) we want". I
think we should constrain it to have the same meaning as elsewhere in
Sirepo. I think this can be resolved by passing run_dir down from the
top-level to all functions that need it. In addition, using the
sim_db_file api will mean we can put files back to the server that other
computes will need and then when we are in that compute the file will
just be in our run_dir (same as other files like lib files).