Fix #3732: Search path for resources, sim_data, pkcli, and templates outside of Sirepo
Resources outside of Sirepo are searched first. This allows a file with the same name
(ex landing.html) to override the Sirepo version.

It is assumed that the files in the other module are organized in the same way as Sirepo. For example,
sim_data for srw is in sirepo.sim_data.srw. For another code it would be in <package_name>.sim_data.<code_name>.

There is a dependent PR in pykern
https://github.com/radiasoft/pykern/pull/104