Fix #3909: Raydata run multiple analyses
Add the 'analysisStatusPanel' to Raydata which shows the status of
selected analysis jobs and allows one to start analysis for missing
jobs. In addition, each row in the table can be clicked if the job is
running or completed. A modal will appear with the png's output from
the notebook.

Two new API's:
- runMulti: Run multiple requests in one API call. In addition, these
requests can be "fire and forget". Normall runSimulation blocks until
the simulation starts. This can take a long while if there is resource
contention. This allows starting the job, not waiting for the
response, and then polling runStatus for updates.
- analysisJob: Call template code on the agent with access to the run
dir.