Need to check for full filepath in sim_data.lib_file_resource_path
Just checking for the existence of a lib dir and then joining the basename
is not enough.

For example, in the case of raydata both the raydata package and the
sirepo package have a lib dir. If we just check for the existence of
the lib dir then we may return the lib dir of the wrong package (the
package that doesn't contain the basename we are looking for). This
change makes it so we check all packages for the file until it is
found or raise if it is not found.