Fix #4038: Always call python parameters.py from slurm
In #4031 the code was changed to get the cmd output by prepare_simulation
and use it to initiate sirepo from slurm (ex srun sirepo srw run_background /path/to/run).

This doesn't work at NERSC because the sirepo command is not installed in the
user's environment. It is only avaiable once we are inside shifter.

So, we reverted to the old behavior of calling `python parameters.py`. But,
with the addition that all prologue/epilogue after calling a code (ex deleting large
intermediate files in srw coherentModesAnimation) must be done in the parameters.py.
So, if a code is going to be run at NERSC there can't be a pkcli run_background command
for it. This forces us to make sure all prologue/epilogue is done in parameters.py.

This means that any epilogue/prologue code will run in a parallel environment so it
must have proper protections (ex ensuring only one mpi process runs it by calling
`sirepo.mpi.restrict_op_to_first_rank`).