Fix #4061: sirepo command not found at NERSC
At NERSC we start the job_agent by supplying the sirepo command
through env (sirepo.job_driver.sbatch.sirepo_cmd). This means that
sirepo commands called from the job agent do not have the proper env.
This was causing the sim_type pkcli lookup to fail.

Modify so the assert for no `run_background` happens at server init.

This has the side effect that all sim_type pkcli's that support sbatch
must be able to be imported without being fully initialized in the
server. This is because we must assert for no run background in the
sim type initialization code (srschema.validate).