Fix #2431: reload supervisor and flask on file changes
Tornado's debug mode causes the server to reload whenever an imported
module changes. But, modules that aren't imported in the supervisor
(ex templates) should cause the system to reload so we pick up the
changes. When the supervisor reloads agent's are dropped and killed on
restart which forces a reload for them and any downstream processes
(ex job_cmd's).

Flask's reloader works like Tornado (reloads when imported python
module changes). But, for extra files it is only relevant to reload
flask when json files changes (ex a code's schema.json).