Fix #4245: set multiElectronNumberOfIterations to 25 in slurm and 5 everywhere else (#4296)
The number of iterations between saves is n(log(n)) where n is
multiElectronNumberOfIterations. This is because the primary process
waits for all worker processes to complete an iteration and counts
that as iteration.

The cause of #4245 was since the number of cores was set to 512 a save
only occured after 262K macro electrons. That resulted in very slow
progress for the simulation which meant no report was requested from
the GUI and so the user interface looked "stuck".