Fix #4583: fix refresh_user loop and add default_proprietary_sim_types back in
default_proprietary_sim_types:
This was removed in https://git.radiasoft.org/sirepo/pulls/4211 . Add it back in
to keep backwards compatibility.

refresh_user:
sirepo.jupyterhub.SirepoAuthenticator.refresh_user should not redirect. Doing so causes a redirect
loop. In the event that a user is not authenticated return False which is what the jupyterhub
protocol expects.