Fix #2447: CancelledError inherits from BaseException on > py 3.8 (#5318)
Change the code to:
- Catch CancelledError explicitly when needed in cases where we were relying on
  it to be a subclass of Exception previously.
- Remove any code that does nothing but raise on CancelledError. This pattern
  was used to not smother CancelledError in cases where we had an `except
  Exception`. Those will no longer smother the error.
- Remove catching of concurrent.futures.base.CancelledError. Tornado never raises
  this error so our best guess is that we don't actually need to catch it.
- Catch BaseException in tornado.Queue. *ANY* type of exception we want context
  for and we always raise at the end so BaseException helps us get more context
  for more types of errors.

In addition, in sirepo.job_supervisro.run re-raise the CancelledError
and remove check for return of False from _send_op. This has no actual
behavior change. We don't await _run so it doesn't matter if the CancelledError
makes it out. But, it is more "correct" to not smother CancelledError and
have it propagate all the way out so a cancel happens fully.