Fix #6305: create service for allocating global resources.
The viz3d app will run an instance of the risivz server and NVIDIA
IndeX for each "simulation" a user starts.  Each of those services
need a set of resources (ip addresses and ports) to be configured
properly. In addition, the resources need to be shared between the
services and GUI so the GUI can make requests to the services.

The resources need to be unique across users so we don't run into
conflicts. The global resource system creates a mechanism for both the
GUI and job agents to request resources.

The resources are allocated per simulation id. This means that within
a simulation the GUI and the job agent have the same resources. This
allows setting up things like having the GUI connect to NVIDIA IndeX
running on a specific port. This also allows a user to have multiple
independent simulations running so they can do things like view
multiple datasets.

There are still some TODO's left to make the systme fully secure and
performant.

In the short-term we need some form of security for the public ports
exposed. This will be a middleware layer that takes a request to a
port, performs auth checks, and then proxies to the correct service. I
plan on investigating a service like configurable-http-proxy (what
jupyterhub uses) or possibly traefik or haproxy. They allow dynamic
api based configuration of a proxy. The request from the GUI may not
go to a specific port but just to sirepo and then be proxied to a
specific local port. TBD when I implement.

In the long-term we need a better mechanism for deciding what agents
need what resources and when. Currently, if the supervisor is
configured with the global resource system
(SIREPO_FEATURE_CONFIG_ENABLE_GLOBAL_RESOURCES=1) then all agents will
have the ability to request resources.  There is state that must be
stored for each agent to keep track of this. For agents that never
request resources this is unnecessary. In addition, we need a way to
garbage collect resources. When an agent is terminated its resources
should be freed so they can be used by another agent. Currently the
viz3d app will have few enough users so we can over-provision
resources so that there are enough resources to go around for all
possible users.