Fix #6378: Get flash oauth working again
This is a partial fix to get flash oauth working again. There are two
aspects that were/are broken:
1. When setting up the tornado _SRequest http_request_uri only added
the path to the protocol and host. In the case of flash this was
/sim-oauth-flash-authorized. But, this missed all of the query params
(?code=xxx). So, authlib would raise a
"MissingCodeException". Changing to full_url() fixes this.
2. Oauth wasn't working in dev because websockets and our oauth flow
are incompatible. We do Set-Cookie when returning the oauth flow
redirect but I believe the redirect response is in a websocket so
Set-Cookie has no impact.

For now, point 2 isn't a problem on prod. We only use websockets in
dev and alpha. So, I want to push this fix through while a figure out
a fix for websockets.