Fix #6542: lock db_upgrade (#6546)
In instances where multiple server process are started we need to lock
db_upgrade so only one is running at a time. Otherwise, they can cause
collisions. For example, in the case of flash files were being created
by one process, deleted by another, and then when the original process
went to use the file it was gone.