for #6642 code improvements
Moved selectedPlots from a saved model value to a plot variable. This is really just local plot state, and only persistent values should end up on the models (and also be present in the model schema).

The existing plotVisibility code was confusing becuase it used "false" as visible, and was also keying off the model name which is static for a plot instance. I changed it so plotVisibility[i] is true if the plot is visible.

$scope.plotLabels() should have been a normal function, rather than on $scope because it is used internally and not referenced from templated html.

$scope.plotLabels() could get replaced with the existing getPlotLabels().

The code now resets plotVisibility() if the current plot labels don't match selectedPlotLabels.

Replaced JSON.stringify() comparison with appState.deepEquals().