Fix #6981: fix computeJobSerial mismatch when doing sbatch login (#7009)
__db_restore was too much. It reset the computeJobSerial which meant
the supervisor and gui were out of sync. Instead just set the
status to MISSING. This let's the runStatus come in, pick up
_sr_exception_in_run and the the user will be prompted to log in.

The job status could be more descriptive. But, MISSING is
reasonable because the user clicked run which means they no longer
care about the existing run simulation status.