Fix #7075: remove refresh_user (#7076)
refresh_user isn't needed for our use. When a user's authentication
status changes we remove the role and remove their server. If they
try to login again the authenticate() code is called.

refresh_user was failing during notifying the hub of activity
after some minutes of uptime. This resulted in many unnecessary
requests and log messages.