Fix #7541: Enable jupyter explicilty (not through sim type)
jupyter is independent of prod/non-prod codes. We either want it
enabled or not and we don't really care that it works by using a
sim type. So, move the configuration into an explicit variable.

On sirepo alpha this will allow us to keep the existing configuration
(not specifying sim types) while moving jupyterhublogin from a moderated_sim_types
to just a sim_type (sirepo is now moderated across the board).jj